move dependencies to a temporary folder on Windows (fixes #228)
When building a self-contained document, Pandoc fetches the content of
HTML dependencies so that they may be inlined; that is, <script
src="foo.js"> is replaced with <script src="data:...">.

In pandoc 1.13 and later, Windows absolute paths beginning with a volume
name are no longer supported when fetching content; that is, <script
src="\Users\bob\foo.js"> is permissible but <script
src="C:\Users\bob\foo.js"> is not.

Because it's possible for a document to reference dependencies from
multiple volumes and there's no format Pandoc accepts on Windows (as far
as I can tell) for a path to another volume, this change works around the
problem by making a temporary copy of the dependencies on Windows, so that
the path can be relative. It's effectively what's already done when
rendering a non self-contained document, but cleaned up with other
intermediates when self-contained.