make the params list read-only
This facilitates integration with tools that execute a single Rmd chunk at a time. With the read-only constriant we can safely add the default params values to the global environment before execution of any chunk in the document (without it we'd have to worry about mutations to the params list in previous chunks).