don't use regexp when replacing img_src
`image_relative()` was using `sub()` to fix an image `src` URL but forgot the `fixed = TRUE` parameter to force the use of simple fixed string substitution.

Instead, this causes regular expression substitution to be used, which leads to all sorts of weird bugs in Windows, which uses backslashes in paths.

One particular case of note is when your image path has a directory containing only numbers in its name, this will cause a failure due to `invalid regular expression: invalid backreference` or something of that sort.