image_relative() must use normalized replacement path
Just using fixed = TRUE does not completely fix the non-self-contained HTML story for images in Windows.

Since the replacement path being used was not normalized (i.e., it was constructed with `relative_to()` instead of `normalized_relative_to()`), the path being generated had Windows-style backslashes and what not, which translated to an incorrect path.