set.seed(100) for default_site
In the websites we've created with rmarkdown to date we've included set.seed(100) in the website Makefile to ensure that html output doesn't constantly churn based on the random number seed (this ends up creating much messier and hard to reason about diffs).

We also include that in our rmarkdown-webstie example: https://github.com/rstudio/rmarkdown-website/blob/master/Makefile#L9

Is there any reason we can think of not to do this here?