Handles soft breaks. Addresses issue #661
Soft-breaks are added by pandoc when a new-line is found but no real
break is required, e.g. when wrapping long lines in lists.
If not handled they are rendered to empty strings, with the result of
merging the last word of a line with the first of the next one.