Do not ignore the knit_meta passed to render()
Currently knit_meta is always overwritten, and this causes a problem in bookdown: when using the Knit-and-Merge approach, knit meta data is collected from each chapter, merged, and finally passed to render(). If it is always overwritten, all meta data will be lost. `knit_meta` is only relevant to documents that need to be knitted, so move these lines back into the if statement.

The breaking change came from ff6019a2048d135dd27427dabe0acaf61aa0c07f