Add option to have arguments NOT shQuoted
This adds logic in pandoc.R/quoted() that checks for presence of ^ at the beginning of an arg and then doesn't shQuote it. It then removes the flagging ^ from the arg before returning the modified args. This is to enable a metadata pandoc arg such as `-M nocite:'@*'` so it doesn't get appended to the command as `"-M nocite:'@*'"` or `-M "nocite:'@*'"`