Prevent clash with KOMA-Script classes (#1509)
* Do not set variable-title by default

This was added in #1284. Setting it by default, however, means that the document cannot be compiled if the document class is set to use KOMA-Script (e.g. `documentclass: scrbook`), or presumably any other class that provides `\subtitle`.

* LaTeX template: Prevent clash with \subtitle

Using `\providecommand` instead of `\newcommand` will prevent an error when using a package that already provides `\subtitle` (as the KOMA-Script classes do).