Fix overlapping content while printing on Firefox with a TOC.
Since there's only one empty left-float, only one page gets pushed to the right.

This is a deeper issue which as far as I can tell is not defined in the
specification, see https://github.com/w3c/csswg-drafts/issues/4434 for an
example with multicol.

This fixes the print issue on Firefox by floating the content to the right,
which produces the desired layout and prevents overlapping the table of
contents.