get rid of this example since it prevents CRAN's macOS machine from buiding the macOS binary
I think this is a bug of CRAN's method to build binaries: at the moment, it doesn't build the binary for rmarkdown because this bookdown example code is broken, but this has been fixed in bookdown and updated on CRAN; however, the current bookdown requires the current version of rmarkdown, then because rmarkdown's binary is not available, bookdown's binary couldn't be built, either; then because bookdown's binary is not available, CRAN refuses to build rmarkdown's binary. In the words, we are stuck in an infinite recursion. The only way to escape is to remove this example first, so CRAN could build the binary for rmarkdown. Then it will build the binary for bookdown.