Specify that `htmltools::htmlPreserve()` should use the pandoc raw attribute rather than preservation tokens when pandoc >= v2.0.
Note that this option will have the intended effect only for versions of htmltools >= 0.5.0.9003.

This causes 1 test to fail b/c it's testing for htmlPreserve output that no longer conforms. The test will pass/fail depending on which version of htmltools is installed, so we may want to make it conditional.