Use Pandoc's default for --email-obfuscation (#1970)
It is the default since Pandoc 1.17.2 so it is a breaking change only for users with a Pandoc version before 1.17.2 and having email adresses in their document. In this case the html source file would have the mailto links obfuscated by default - html view in browser would still be the same.