Allow to add specific shiny-prerendered parts in HTML template (#2064)
A new pandoc variable `shiny-prerendered` is set when rendering a shiny prerendered document. This allow to add specific content for this type of rendered document in an HTML template. 
It is used in all HTML templates to add a special comment `<!-- HEAD_CONTENT -->` required by htmltools / shiny to reinsert the HTML dependencies. Following #1942, it was put by default at the end of the error which caused issue in leanr, e.g rstudio/learnr#499

This will fix #2060 following change in 49c0c2b.