Support all math rendering engines from Pandoc in HTML documents (#1940)
Pandoc supports different math method for HTML document (https://pandoc.org/MANUAL.html#math-rendering-in-html) that we now support in HTML document through `math_method` argument.  

`mathjax` will still take precedence though if `"default"` is not used for backward compatibility. Support should be available in all format based on those in **rmarkdown**. `math_method = "katex"` will be included as a HTML dependency so that we can evade the downloading and encoding process of Pandoc for self contained mode. This feature has a minimum requirement of Pandoc 2.0 as it requires `data-external = 1`. 

Co-authored-by: Christophe Dervieux <cderv@rstudio.com>