add shiny_prerendered special comment via pandoc_include_args (#2249)
Special comment `<!-- HEAD_CONTENT -->` is now added in `$header-includes$` and there is not need anymore to add it in template. This will make it work with any custom template. As a reminder, this comments is used by htmltools to insert head content - which shiny does for html dependencies found during knitting. They will now be inserted at the same place as with non shiny document.

This offers a better fix than #2064 that aimed to be an improvement over #1942 to fix #1912. #2064 workaround is still kept though as safety measure but should not be needed

Co-authored-by: Christophe Dervieux <christophe.dervieux@gmail.com>