Vectorise str_replace_all() (#2416)
In the development version of stringr, I'm exploring vectorising the function form of `str_replace_all()` (https://github.com/tidyverse/stringr/pull/4620) because it offers considerable performance improvements. This patch quickly hacks rmarkdown to work with this change. It would probably be worthwhile to vectorise the two users of `process_html_res()` as that is likely to give a small, but meaningful, performance boost when there are many matches.